`IndexRowBuilder` rework complete.
Now uses Ox under the hood, rather than Nokogiri. :clap:

Pinging Issue #265.

RSpec: 1048 examples, 0 failures, 1 pending
RuboCop: 299 files inspected, no offenses detected
Coverage: 3176 / 3189 LOC (99.59%) covered.
